feat: make driver to be more flexible and have modularity
