+++ examples/basic.ts
@@ -1,27 +1,6 @@
 import { driver } from '../mod.ts'
 
-driver({ browser: 'chrome' })
-	.then(({ service }) => {
-		service.cases = [
-			{
-				name: 'Verify Failed Title',
-				fn: async ({ builder, assert }) => {
-					const title = await builder.getTitle()
-					assert.assertEquals(title, 'Drowsers')
-				},
-			},
-			{
-				name: 'Verify Title',
-				fn: async ({ builder, assert }) => {
-					const title = await builder.getTitle()
-					assert.assertEquals(title, 'Drowser')
-				},
-			},
-		]
-	})
-	.catch((error) => console.log(error))
-
-driver({ browser: 'firefox' })
+driver({ browser: 'safari' })
 	.then(({ service }) => {
 		service.cases = [
 			{
